requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
gradio_client>=0.0.1
gtts>=2.3.0
discord.py>=2.3.0
